커뮤니티 css / 미비된 부분 수정!
