--- src/main/java/com/kh/sinsa/community/controller/CommunityOotdListServlet.java
@@ -1,25 +1,0 @@
-package com.kh.sinsa.community.controller;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/**
- * Servlet implementation class CommunityOotdListServlet
- */
-@WebServlet("/community/communityOotdList")
-public class CommunityOotdListServlet extends HttpServlet {
-	private static final long serialVersionUID = 1L;
-
-	/**
-	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
-	 */
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		request.getRequestDispatcher("/WEB-INF/views/ootd_community/ootdList.jsp")
-		.forward(request, response);
-	}
-
-}
