+++ src/main/java/com/kh/sinsa/community/controller/codi/CodiCommentAdd.java
@@ -1,7 +1,6 @@
-package com.kh.sinsa.community.controller;
+package com.kh.sinsa.community.controller.codi;
 
 import java.io.IOException;
-
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -13,10 +12,10 @@ import com.kh.sinsa.community.model.dto.CommunityComment;
 import com.kh.sinsa.community.model.service.CommunityService;
 
 /**
- * Servlet implementation class CommunityCommentAddServlet
+ * Servlet implementation class CodiCommentAdd
  */
-@WebServlet("/share/shareCommentAdd")
-public class CommunityCommentAddServlet extends HttpServlet {
+@WebServlet("/community/codiCommentAdd")
+public class CodiCommentAdd extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 	private CommunityService communityService = new CommunityService();
 
@@ -24,27 +23,20 @@ public class CommunityCommentAddServlet extends HttpServlet {
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int result = 0;
-		
 		try {
 			// 1. 사용자 입력값 처리
 			String commNo = request.getParameter("commNo");
-			String writer = request.getParameter("writer");
-			String content = request.getParameter("content");
-			CommentLevel commentLevel = CommentLevel.valueOf(Integer.parseInt(request.getParameter("commentLevel")));
-			String commentRef = request.getParameter("commentRef");
+			String writer = request.getParameter("userId");
+			String content = request.getParameter("commCommentContent");
+			CommentLevel commentLevel = CommentLevel.valueOf(Integer.parseInt(request.getParameter("commCommentLevel")));
+			String commentRef = request.getParameter("commCommentRef");
 			CommunityComment communityComment = new CommunityComment(null, commNo, writer, content, null, commentLevel, commentRef);
 			
 			// 2. 업무로직
+			int result = communityService.codiCommentAdd(communityComment);
 
-			result = communityService.insertCommunityComment(communityComment);
-			System.out.println("ccccc : " + communityComment);
-
-
-			
 			// 3. redirect 응답
-			response.sendRedirect(request.getContextPath() + "/share/shareView?no=" + commNo);
-			
+			response.sendRedirect(request.getHeader("Referer"));
 			
 		}
 		catch(Exception e) {
