+++ src/main/java/com/kh/sinsa/community/controller/share/CommunityCommentDeleteServlet.java
@@ -1,4 +1,4 @@
-package com.kh.sinsa.community.controller;
+package com.kh.sinsa.community.controller.share;
 
 import java.io.IOException;
