+++ src/main/java/com/kh/sinsa/community/controller/share/CommunityViewServlet.java
@@ -1,4 +1,4 @@
-package com.kh.sinsa.community.controller;
+package com.kh.sinsa.community.controller.share;
 
 import java.io.IOException;
 import java.util.List;
