+++ src/main/java/com/kh/sinsa/community/model/dao/CommunityDao.java
@@ -1492,13 +1492,6 @@ public class CommunityDao {
 		}
 		return list;
 	}
-
-
-
-
-
-
-
-
+	
 
 }
