+++ src/main/resources/sql/community/community-query.properties
@@ -80,9 +80,3 @@ talkAlign = select * from (select row_number () over (order by # desc)rnum, c.* 
 getTalkTotalContent = select count(*) from community where comm_no like 'C50%'
 talkSearch = select * from (select row_number () over (order by comm_date desc)rnum, c.* from community c where comm_no like 'C50%' and comm_title like ?) c where rnum between ? and ?
 
-
-
-
-
-
-
