+++ src/main/webapp/WEB-INF/views/free/freeAdd.jsp
@@ -8,27 +8,17 @@
 	<div id="container">
 		<div id="header1"></div>
 		<div id="header2">COMMUNITY</div>
-		<nav class="navi">
-			<ul>
-				<li><a href="#">OOTD</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/codiList">코디북</a></li>
-				<li><a href="<%=request.getContextPath()%>/share/shareList">정보공유</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/freeList">자유게시판</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/talkList">패션토크</a></li>
-			</ul>
-		</nav>
+	
 		<form name="freeAddFrm"
 			action="<%=request.getContextPath()%>/community/freeAdd" 
 			method="post"
 			encType = "multipart/form-data">
-			<h2 style="font-size: 30px; color: black;">패션토크 게시글 작성</h2>
+			<h2 style="font-size: 30px; color: black;">자유게시판 게시글 작성</h2>
 			<hr style="border-top: 3px solid black;">
 			<div id="content" style="margin: 30px;">
 				<span class="drop"> 
 				<select name="one" id="drop-select" onchange="if(this.value) location.href=(this.value)">
 						<option class="option" value="">게시판 선택</option>
-						<option class="option" value="#">OOTD</option>
-						<option class="option" value="#">코디북</option>
 						<option class="option" value="<%=request.getContextPath()%>/share/shareAdd">정보공유</option>
 						<option class="option" value="<%= request.getContextPath() %>/community/freeAdd" selected>자유게시판</option>
 						<option class="option" value="<%= request.getContextPath() %>/community/talkAdd"/>패션토크</option>
