+++ src/main/webapp/WEB-INF/views/free/freeEdit.jsp
@@ -17,33 +17,15 @@
 	<div id="container">
 		<div id="header1"></div>
 		<div id="header2">COMMUNITY</div>
-		<nav class="navi">
-			<ul>
-				<li><a href="#">OOTD</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/codiList">코디북</a></li>
-				<li><a href="<%= request.getContextPath() %>/share/shareList">정보공유</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/freeList">자유게시판</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/talkList">">패션토크</a></li>
-			</ul>
-		</nav>
+		
 		<form name="freeEditFrm"
 			action="<%=request.getContextPath()%>/community/freeEdit" 
 			enctype="multipart/form-data"
 			method="post">
-			<h2 style="font-size: 30px; color: black;">패션토크 게시글 수정</h2>
-			<hr style="border-top: 3px solid black;">
+			<h2 style="font-size: 30px; color: purple; font-weight:bold;">자유게시판 게시글 수정</h2>
+			<hr style="border-top: 3px solid purple;">
 			<div id="content" style="margin: 30px;">
 				
-				<span class="dropdown"> 
-				<select onchange="if(this.value) location.href=(this.value);" name="one" class="dropdown-select">
-						<option value="0">게시판 선택</option>
-						<option value="1">OOTD</option>
-						<option value="<%= request.getContextPath() %>/community/codiEdit">코디북</option>
-						<option value="<%=request.getContextPath()%>/share/shareEdit" selected />정보공유</option>
-						<option value="<%= request.getContextPath() %>/community/freeEdit" >자유게시판</option>
-						<option value="<%= request.getContextPath() %>/community/talkEdit">">패션토크</option>
-				</select>
-				</span> 
 				<br> 
 
 				<input type="hidden" name="no" value="<%= community.getCommNo() %>">
@@ -72,7 +54,7 @@
 				<input 
 					type="checkbox" name="delFile" 
 					id="delFile<%= i %>" value="<%= attach.getNo() %>"/>
-				<label for="delFile<%= i %>">삭제</label>>
+				<label for="delFile<%= i %>">삭제</label>
 				
 				
 				<%
