--- src/main/webapp/WEB-INF/views/ootd_community/ootdView.jsp
@@ -1,6 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<%@ include file="/WEB-INF/views/common/header.jsp"%>
-<link rel="stylesheet" href="<%=request.getContextPath()%>/css/ootdView.css" /> 
- 
-<%@ include file="/WEB-INF/views/common/footer.jsp"%>
