+++ src/main/webapp/WEB-INF/views/share_community/shareList.jsp
@@ -12,22 +12,18 @@
 %>
 <link rel="stylesheet" href="<%=request.getContextPath()%>/css/share/shareList.css" />
         <div class="container">
-            <div id="headlist">
+            
                 <div id="header2">COMMUNITY</div>
-                <nav class="navi">
-                  <ul>
-                      <li><a href="#">OOTD</a></li>
-                      <li><a href="<%= request.getContextPath() %>/community/codiList">코디북</a></li>
-                      <li><a href="<%=request.getContextPath()%>/share/shareList">정보공유</a></li>
-                      <li><a href="<%= request.getContextPath() %>/community/freeist">자유게시판</a></li>
-                      <li><a href="<%= request.getContextPath() %>/community/talkList">패션토크</a></li>
-                  </ul>
-              </nav>
+                <nav class="py-2 bg-white border-top border-bottom" id="commnavi">
+	          <ul class="nav me-auto">
+	            <li class="nav-item"><a href="<%= request.getContextPath() %>/community/codiList" class="nav-link link-dark px-2">코디북</a></li>
+	            <li class="nav-item"><a href="<%= request.getContextPath() %>/share/shareList" class="nav-link link-dark px-2">정보공유</a></li>
+	            <li class="nav-item"><a href="<%= request.getContextPath() %>/community/freeList" class="nav-link link-dark px-2">자유게시판</a></li>
+	            <li class="nav-item"><a href="<%= request.getContextPath() %>/community/talkList" class="nav-link link-dark px-2">패션토크</a></li>
+	          </ul>
+	      </nav>
         
-        
-
-            </div>
-
+			<br><br><br>
 
             <div id="main">
                 <h1>정보공유 게시판<span style="font-size: 20px;"> Share Information</span></h1>
@@ -112,7 +108,9 @@
                 <input  style="float:right;"
 						type="button" value="글쓰기" id="write"
 						onclick="location.href='<%= request.getContextPath() %>/share/shareAdd';"/>
+						<br>
  			 </div>
+ 			 
             </div>
         
         <script>
