+++ src/main/webapp/WEB-INF/views/talk/talkAdd.jsp
@@ -8,15 +8,7 @@
 	<div id="container">
 		<div id="header1"></div>
 		<div id="header2">COMMUNITY</div>
-		<nav class="navi">
-			<ul>
-				<li><a href="#">OOTD</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/codiList">코디북</a></li>
-				<li><a href="<%=request.getContextPath()%>/share/shareList">정보공유</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/freeList">자유게시판</a></li>
-				<li><a href="<%= request.getContextPath() %>/community/talkList">패션토크</a></li>
-			</ul>
-		</nav>
+
 		<form name="talkAddFrm"
 			action="<%=request.getContextPath()%>/community/talkAdd" 
 			method="post"
@@ -28,7 +20,6 @@
 				<select name="one" id="drop-select" onchange="if(this.value) location.href=(this.value)">
 						<option class="option" value="">게시판 선택</option>
 						<option class="option" value="#">OOTD</option>
-						<option class="option" value="#">코디북</option>
 						<option class="option" value="<%=request.getContextPath()%>/share/shareAdd">정보공유</option>
 						<option class="option" value="<%= request.getContextPath() %>/community/freeAdd">자유게시판</option>
 						<option class="option" value="<%= request.getContextPath() %>/community/talkAdd" selected />패션토크</option>
