+++ src/main/webapp/css/notice/noticeList.css
@@ -1,0 +1,105 @@
+@charset "EUC-KR";
+
+#header2 {margin: 100px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
+      .navi ul {display: flex; justify-content:center;}
+      .navi ul li a {text-decoration:none;color:rgb(131, 130, 130); padding:0 30px; font-weight: bold; font-size: 20px;}
+      .navi ul li:not(:last-of-type)::after {content: "|"}    
+      .navi ul li a:hover    {color:rgb(0, 0, 0);}
+
+/* 홈페이지 규격 */
+.container {
+    width: 1024px;
+    margin: 0 auto;  /* 화면에 가운데 정렬하기 위함 */
+}
+
+/* header */
+
+#logo {
+    margin: 20px auto 0;
+}
+#logo > img {
+    width: 150px;
+    height: auto;
+}
+.menubar {
+    color: rgb(50, 50, 50);
+    font-size: 18px;
+    font-weight: 600;
+}
+
+/* 컨텐츠 영역 */
+
+#main {
+    width: 100%;
+    height: auto;
+
+} 
+
+.align{
+	width : 150px;
+	float : left;
+}
+
+.search{
+	width : 250px;
+	float : right;
+}
+
+div#list{
+  
+    height: auto;
+}
+#write{
+    width: 60px;
+    background-color: black;
+    color: white;
+    font-weight: bold;
+}
+
+#option{
+    display: flex;
+    justify-content: space-between;
+    margin-left: 30px;
+    margin-right: 30px;
+    margin-bottom: 50px;
+    
+}
+
+#standard{
+    background-color: gray;  
+    border-radius: 5px;
+    height: 40px;
+}
+
+#tbl-board{
+    
+    width: 800px;
+    border: 1px solid black;
+    border-collapse: collapse;
+    margin: 0 auto;
+
+}
+
+th, td{
+    border: 1px solid black;
+    text-align: center;
+   
+}
+#boardTitle{
+    color: rgb(50, 50, 50);
+}
+
+div#pagebar{
+			margin-top:20px;
+			text-align:center; 
+			background-color:gray; 
+}
+div#pagebar span.cPage{
+			color: white; 
+			margin-right: 5px;
+}
+			
+div#pagebar a{
+			margin-right: 5px;
+}
+
