+++ src/main/webapp/css/notice/noticeView.css
@@ -1,27 +1,10 @@
 @charset "EUC-KR";
 
-#header2 {margin: 20px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
-           
-.nav{
-    justify-content: space-around;
-}
-
-#gita{
-	text-align: right;
-	margin-top: 30px;
-	margin-right: 30px;
-	font-size: 13px;
-}
-
-.justtext p{
-	font-size: 15px;
-	line-height: 5px;
-}
-
-#warn{
-	font-size: 12px;
-	color: red;
-}
+#header2 {margin: 100px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
+      .navi ul {display: flex; justify-content:center;}
+      .navi ul li a {text-decoration:none;color:rgb(131, 130, 130); padding:0 30px; font-weight: bold; font-size: 20px;}
+      .navi ul li:not(:last-of-type)::after {content: "|"}    
+      .navi ul li a:hover    {color:rgb(0, 0, 0);}
 
 /* 홈페이지 규격 */
 .container {
@@ -35,20 +18,36 @@
     margin: 0 auto;  /* 화면에 가운데 정렬하기 위함 */
 }
 
-#readRecomm{
-	float : right;
-}
+/* header */
 
-#editDel{
-	float : right;
+#logo {
+    margin: 20px auto 0;
 }
-
+#logo > img {
+    width: 150px;
+    height: auto;
+}
+a {
+    color: rgb(50, 50, 50);
+    font-size: 18px;
+    font-weight: 600;
+}
 
 /* 컨텐츠 영역 */
 .main {
     width: 1024px;
     height: auto;;
 
+}
+
+#goList{
+	float: left;	
+}
+
+#btn_goList{
+	border-radius:30%;
+	background-color:white;
+	
 }
