--- src/main/webapp/css/ootd_community/ootdAdd.css
@@ -1,102 +1,0 @@
-@charset "UTF-8";
-      body{max-width: 1024px; margin: 0 auto;}
-        li    {list-style:none; }
-      #header2 {margin: 100px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
-      #content { text-align: center;}
-      #file {float: left; margin-left: 80px;}
-        nav    {transform: translateY(-10px); }
-        nav ul {display: flex; justify-content:center;}
-        nav ul li a {text-decoration:none;color:rgb(131, 130, 130); padding:0 30px; font-weight: bold; font-size: 20px;}
-        nav ul li:not(:last-of-type)::after {content: "|"}    
-        nav ul li a:hover    {color:rgb(0, 0, 0);}
-      img {width: 353px; height: 235px;}
-        
-      .title {
-        position: relative;
-        width: 725px;
-        display: flex;
-        margin-left: 150px;
-      }
-      #title {
-          width: 100%;
-          border-radius: 10px;
-          border: 1px solid #bbb;
-        
-          padding: 10px 12px;
-          display: flex;
-      }
-
-      .dropdown {
-        
-        margin: 0 150px;
-        display: flex;
-        position: relative;
-        overflow: hidden;
-        height: 25px;
-        width: 150px;
-        background: #f2f2f2;
-        border: 1px solid;
-        border-color: white #f7f7f7 #f5f5f5;
-        border-radius: 3px;
-        background-image: -webkit-linear-gradient(top, transparent, rgba(0, 0, 0, 0.06));
-        background-image: -moz-linear-gradient(top, transparent, rgba(0, 0, 0, 0.06));
-        background-image: -o-linear-gradient(top, transparent, rgba(0, 0, 0, 0.06));
-        background-image: linear-gradient(to bottom, transparent, rgba(0, 0, 0, 0.06));
-        -webkit-box-shadow: 0 1px 1px rgba(0, 0, 0, 0.08);
-        box-shadow: 0 1px 1px rgba(0, 0, 0, 0.08);
-        
-        
-      }
-      .dropdown:before, .dropdown:after {
-        content: '';
-        position: absolute;
-        z-index: 2;
-        top: 9px;
-        right: 10px;
-        width: 0;
-        height: 0;
-        border: 4px dashed;
-        border-color: #888 transparent;
-        pointer-events: none;
-      }
-      .dropdown:before {
-        border-bottom-style: solid;
-        border-top: none;
-      }
-      .dropdown:after {
-        margin-top: 7px;
-        border-top-style: solid;
-        border-bottom: none;
-      }
-
-      .dropdown-select {
-        position: relative;
-        width: 130%;
-        margin: 0;
-        padding: 6px 8px 6px 10px;
-        height: 28px;
-        line-height: 14px;
-        font-size: 12px;
-        color: #62717a;
-        text-shadow: 0 1px white;
-        background: #f2f2f2;
-        background: rgba(0, 0, 0, 0);
-        border: 0;
-        border-radius: 0;
-        -webkit-appearance: none;
-      }
-      .dropdown-select:focus {
-        z-index: 3;
-        width: 100%;
-        color: #394349;
-        outline: 2px solid #49aff2;
-        outline-offset: -2px;
-      }
-      .dropdown-select > option {
-        margin: 3px;
-        padding: 6px 8px;
-        text-shadow: none;
-        background: #f2f2f2;
-        border-radius: 3px;
-        cursor: pointer;
-      }
