--- src/main/webapp/css/ootd_community/ootdEdit.css
@@ -1,1 +1,0 @@
-@charset "UTF-8";
