+++ src/main/webapp/css/share/shareList.css
@@ -1,10 +1,31 @@
 @charset "EUC-KR";
 
-#header2 {margin: 100px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
+#header2 {margin: 20px 0 50px 0;  text-align: center; font-weight: bold; font-size: 30px;}
       .navi ul {display: flex; justify-content:center;}
       .navi ul li a {text-decoration:none;color:rgb(131, 130, 130); padding:0 30px; font-weight: bold; font-size: 20px;}
       .navi ul li:not(:last-of-type)::after {content: "|"}    
       .navi ul li a:hover    {color:rgb(0, 0, 0);}
+      
+.nav{
+    justify-content: space-around;
+}
+
+#gita{
+	text-align: right;
+	margin-top: 30px;
+	margin-right: 30px;
+	font-size: 13px;
+}
+
+.justtext p{
+	font-size: 15px;
+	line-height: 5px;
+}
+
+#warn{
+	font-size: 12px;
+	color: red;
+}
 
 /* 홈페이지 규격 */
 .container {
@@ -73,17 +94,30 @@ div#list{
 
 #tbl-board{
     
-    width: 800px;
+    width: 100%;
     border: 1px solid black;
     border-collapse: collapse;
     margin: 0 auto;
+    border-left:none;
+    border-right:none;
 
 }
 
+th{
+	background-color : #DFF6FF;
+}
+
+
 th, td{
     border: 1px solid black;
+    border-left:none;
+    border-right:none;
     text-align: center;
+    padding : 5px 0;
    
+}
+td a {
+	color : black;
 }
 #boardTitle{
     color: rgb(50, 50, 50);
@@ -92,10 +126,10 @@ th, td{
 div#pagebar{
 			margin-top:20px;
 			text-align:center; 
-			background-color:gray; 
+			background-color:white; 
 }
 div#pagebar span.cPage{
-			color: white; 
+			color: black; 
 			margin-right: 5px;
 }
